move CUserSettings to different files
